bruker usefromcontext på resten av tilleggssporsmal
