+++ src/components/sporsmal/tilleggssporsmal/OpplysningeneErFeil.tsx
@@ -1,17 +1,15 @@
 import React from 'react';
-import { FieldError } from 'react-hook-form/dist/types';
 import { Fieldset, SkjemaGruppe, Checkbox } from 'nav-frontend-skjema';
 import tekster from '../sporsmal-tekster';
 import { AlertStripeHjelper } from '../../../utils/alertstripe-utils';
 import { Skjemafelt } from '../../../types/sporsmalTypes';
+import { useFormContext } from 'react-hook-form';
 
 interface OpplysningeneErFeilProps {
     vis: boolean;
     visAlertstripeAvbryt: boolean;
     visAlertstripeBrukArbeidsgiver: boolean;
     visAlertstripeBruk: boolean;
-    register: any;
-    errors: Partial<Record<string, FieldError>>;
 }
 
 const OpplysningeneErFeil = ({
@@ -19,9 +17,9 @@ const OpplysningeneErFeil = ({
     visAlertstripeAvbryt,
     visAlertstripeBrukArbeidsgiver,
     visAlertstripeBruk,
-    register,
-    errors,
 }: OpplysningeneErFeilProps) => {
+    const { register, errors} = useFormContext();
+
     if (!vis) {
         return null;
     }
