+++ src/types/sporsmalTypes.ts
@@ -1,3 +1,5 @@
+import { CustomLocale } from "flatpickr/dist/types/locale";
+
 export enum Skjemafelt {
     OPPLYSNINGENE_ER_RIKTIGE = 'opplysningeneErRiktige',
     PERIODE = 'periode',
@@ -25,3 +27,29 @@ export enum Arbeidsforhold {
     ARBEIDSLEDIG = 'arbeidsledig',
     INGENTING_PASSER = 'ingentingPasser',
 }
+
+export const locale: CustomLocale = {
+    rangeSeparator: ' til ',
+    firstDayOfWeek: 1,
+    weekdays: {
+        shorthand: ['søn', 'man', 'tirs', 'ons', 'tors', 'fre', 'lør'],
+        longhand: ['søndag', 'mandag', 'tirsadg', 'onsdag', 'torsdag', 'fredag', 'lørdag'],
+    },
+    months: {
+        shorthand: ['jan', 'feb', 'mar', 'apr', 'mai', 'jun', 'jul', 'aug', 'sep', 'okt', 'nov', 'des'],
+        longhand: [
+            'januar',
+            'februar',
+            'mars',
+            'april',
+            'mai',
+            'juni',
+            'juli',
+            'aug',
+            'september',
+            'oktober',
+            'november',
+            'desember',
+        ],
+    },
+};
